// src/app/model/AddUser.model.ts
export class AddUserModel{
    nombre?: string;
	appPaterno?: string;
	appMaterno?: string;
	correo?: string;
    correoConfirm?: string;
	pass?: string;
	passConfirm?: string;
    celular?: string;
    idPais?: number;
    idRegion?: number;
    idComuna?: number;
    direccion?: string;
    genero?: number;

    constructor(){
      
    }
   
}